requirements.txt
# Core FastAPI and web framework
fastapi>=0.104.1
uvicorn[standard]>=0.24.0

# Database and ORM
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
alembic>=1.12.1

# Task queue and background jobs
celery>=5.3.4
redis>=5.0.1

# Environment and configuration
python-dotenv>=1.0.0

# API integrations
firecrawl-py>=0.0.16
google-generativeai>=0.3.0

# Data processing and utilities
pydantic>=2.5.0
python-dateutil>=2.8.2
requests>=2.31.0

# Image processing
pillow>=10.1.0

# Development and testing
pytest>=7.4.3
pytest-asyncio>=0.21.1

# Web dashboard
jinja2>=3.1.2

# Optional: For development server
watchfiles>=0.21.0
